backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
anthropic==0.40.0
openai==1.54.0
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
tavily-python==0.5.0
beautifulsoup4==4.12.3
